main/requirements.txt
anyio==4.9.0
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cleo==2.1.0
colorama==0.4.6
comtypes==1.4.10
contourpy==1.3.3
crashtest==0.4.1
cryptography==45.0.6
cycler==0.12.1
distlib==0.3.9
docopt==0.6.2
dulwich==0.22.8
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
fonttools==4.60.1
fsspec==2025.7.0
fuzzywuzzy==0.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
joblib==1.5.1
keyring==25.6.0
kiwisolver==1.4.9
Levenshtein==0.27.1
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.10.6
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.5
numpy==2.2.4
packaging==25.0
pbs-installer==2025.6.6
pdfminer==20191125
pdfminer.six==20250506
pillow==11.3.0
pipreqs==0.4.13
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry==2.1.3
poetry-core==2.1.3
psutil==7.0.0
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.23.0
pylance==0.24.1
pyparsing==3.2.5
PyPDF2==3.0.1
pyproject_hooks==1.2.0
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-frontmatter==1.1.0
python-Levenshtein==0.27.1
pyvda==0.5.0
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.12.2
regex==2025.7.34
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
shellingham==1.5.4
shiboken6==6.9.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.4
tomlkit==0.13.3
torch==2.8.0
tqdm==4.67.1
transformers==4.55.0
trove-classifiers==2025.5.9.12
typing_extensions==4.14.1
urllib3==2.4.0
virtualenv==20.31.2
yarg==0.1.10
zstandard==0.23.0
